requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
asyncpg>=0.27.0
pydantic>=2.0.0
alembic>=1.15.2
python-multipart>=0.0.6
pydantic_settings>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
email-validator>=2.0.0
